update:更新model和dto里的id-》personInd，id-〉spaceId。统一参数类型为Integer

+++ backend/src/main/java/edu/fudan/se/sctap_lowcode_tool/neo4jRepository/PersonNodeRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PersonNodeRepository extends Neo4jRepository<PersonNode, Long> {
+public interface PersonNodeRepository extends Neo4jRepository<PersonNode, Integer> {
     List<PersonNode> findByCurrentSpace(SpaceNode spaceNode);
 }
